separate openrave independent part query_grasp
