+++ cob_grasp_generation/launch/grasp_generation.launch
@@ -5,5 +5,6 @@
   <env name="OPENRAVE_DATA" value="$(find cob_grasp_generation)/files:$(optenv OPENRAVE_DATA)"/>
 
   <node name="cob_grasp_generation_OR" pkg="cob_grasp_generation" type="cob_grasp_generation_action_server.py" respawn="false" output="screen"/>
+  <node name="cob_grasp_query" pkg="cob_grasp_generation" type="cob_grasp_query_action_server.py" respawn="false" output="screen"/>
 
 </launch>
