+++ cob_grasp_generation/scripts/cob_grasp_generation_action_server.py
@@ -20,7 +20,7 @@ import rospy
 import actionlib
 import moveit_msgs.msg
 import cob_grasp_generation.msg
-from cob_grasp_generation import or_grasp_generation
+from cob_grasp_generation import or_grasp_generation, grasp_query_utils
 
 class CobGraspGenerationActionServer(object):
 
@@ -29,8 +29,6 @@ class CobGraspGenerationActionServer(object):
 
     self._as_generate = actionlib.SimpleActionServer('generate_grasps', cob_grasp_generation.msg.GenerateGraspsAction, execute_cb=self.generate_cb, auto_start = False)
     self._as_generate.start()
-    self._as_query = actionlib.SimpleActionServer('query_grasps', cob_grasp_generation.msg.QueryGraspsAction, execute_cb=self.query_cb, auto_start = False)
-    self._as_query.start()
     self._as_show = actionlib.SimpleActionServer('show_grasps', cob_grasp_generation.msg.ShowGraspsAction, execute_cb=self.show_cb, auto_start = False)
     self._as_show.start()
 
@@ -42,7 +40,7 @@ class CobGraspGenerationActionServer(object):
 
     rospy.loginfo('Generating grasps for object %s using gripper_type %s' % (goal.object_name, goal.gripper_type))
 
-    if self.orgg.check_database(goal.object_name, goal.gripper_type):
+    if grasp_query_utils.check_database(goal.object_name, goal.gripper_type):
         rospy.logwarn('Grasps for object %s exist in the database.', goal.object_name)
         success = False
     else:
@@ -64,41 +62,12 @@ class CobGraspGenerationActionServer(object):
       self._as_generate.set_aborted(result)
 
 
-
-  def query_cb(self, goal):
-    success = False
-    grasp_list = []
-
-    rospy.loginfo('Querying grasps for object %s using gripper_type %s' % (goal.object_name, goal.gripper_type))
-
-    if self.orgg.check_database(goal.object_name, goal.gripper_type):
-        rospy.loginfo('GraspTable for object %s exist in the database.', goal.object_name)
-        rospy.loginfo('Returning grasp list for selected object.')
-        grasp_list = self.orgg.get_grasps(goal.object_name, goal.gripper_type, goal.grasp_id, goal.num_grasps, goal.threshold)
-    else:
-        rospy.logwarn('GraspTable for object %s does not exist!',goal.object_name)
-
-    if not (grasp_list == []):
-        success = True
-
-    result   = cob_grasp_generation.msg.QueryGraspsResult()
-    result.success = success
-    result.grasp_list = grasp_list
-
-    if success:
-      rospy.loginfo('Query: Succeeded')
-      self._as_query.set_succeeded(result)
-    else:
-      rospy.logwarn('Query: Failed')
-      self._as_query.set_aborted(result)
-
-
   def show_cb(self, goal):
     success = False
 
     rospy.loginfo('Show grasp %i for object %s using gripper_type %s' % (goal.grasp_id, goal.object_name, goal.gripper_type))
 
-    if self.orgg.check_database(goal.object_name, goal.gripper_type):
+    if grasp_query_utils.check_database(goal.object_name, goal.gripper_type):
       self.orgg.setup_environment(goal.object_name, goal.gripper_type, viewer=True)
       rospy.loginfo('Display Grasp. Object: %s | ID: %i' % (goal.object_name, goal.grasp_id))
       self.orgg.show_grasp(goal.object_name, goal.gripper_type, goal.grasp_id, goal.sort_by_quality)
