+++ cob_grasp_generation/src/cob_grasp_generation/__init__.py
@@ -1,1 +1,1 @@
-from cob_grasp_generation import *
+
