+++ cob_pick_place_action/launch/cob_pick_place.launch
@@ -2,6 +2,6 @@
 <launch>
    <arg name="manipulator_group"/>
 
-   <include file="$(find cob_grasp_generation)/launch/grasp_generation.launch"/>
+   <node name="cob_grasp_query" pkg="cob_grasp_generation" type="cob_grasp_query_action_server.py" respawn="false" output="screen"/>
    <node name="cob_pick_place_action" pkg="cob_pick_place_action" type="cob_pick_place_action" args="$(arg manipulator_group)" respawn="false" output="screen"/>
 </launch>
