+++ cob_pick_place_action/scripts/cob_pick_action_client.py
@@ -73,17 +73,17 @@ def cob_pick_action_client():
 	goal.object_pose.header.stamp = rospy.Time.now()
 	goal.object_pose.header.frame_id = "base_footprint"
 	### cob3
-	goal.object_pose.pose.position.x = random.uniform(-0.8, -0.6)
-	goal.object_pose.pose.position.y = random.uniform(-0.3,  0.3)
-	goal.object_pose.pose.position.z = random.uniform( 0.8,  1.1)
-	goal.object_pose.pose.orientation.x, goal.object_pose.pose.orientation.y, goal.object_pose.pose.orientation.z, goal.object_pose.pose.orientation.w = quaternion_from_euler(random.uniform(-pi/2, pi/2),random.uniform(-pi/2, pi/2),random.uniform(-pi/2, pi/2))
-	goal.gripper_type = "sdh"
+	#goal.object_pose.pose.position.x = random.uniform(-0.8, -0.6)
+	#goal.object_pose.pose.position.y = random.uniform(-0.3,  0.3)
+	#goal.object_pose.pose.position.z = random.uniform( 0.8,  1.1)
+	#goal.object_pose.pose.orientation.x, goal.object_pose.pose.orientation.y, goal.object_pose.pose.orientation.z, goal.object_pose.pose.orientation.w = quaternion_from_euler(random.uniform(-pi/2, pi/2),random.uniform(-pi/2, pi/2),random.uniform(-pi/2, pi/2))
+	#goal.gripper_type = "sdh"
 	### cob4
-	# goal.object_pose.pose.position.x = random.uniform(0.4, 0.6)
-	# goal.object_pose.pose.position.y = random.uniform(0.5, 0.7)
-	# goal.object_pose.pose.position.z = random.uniform(0.8, 1.0)
-	# goal.object_pose.pose.orientation.x, goal.object_pose.pose.orientation.y, goal.object_pose.pose.orientation.z, goal.object_pose.pose.orientation.w = quaternion_from_euler(random.uniform(-pi/2, pi/2),random.uniform(-pi/2, pi/2),random.uniform(-pi/2, pi/2))
-	# goal.gripper_type = "sdhx"
+	goal.object_pose.pose.position.x = random.uniform(0.4, 0.6)
+	goal.object_pose.pose.position.y = random.uniform(0.5, 0.7)
+	goal.object_pose.pose.position.z = random.uniform(0.8, 1.0)
+	goal.object_pose.pose.orientation.x, goal.object_pose.pose.orientation.y, goal.object_pose.pose.orientation.z, goal.object_pose.pose.orientation.w = quaternion_from_euler(random.uniform(-pi/2, pi/2),random.uniform(-pi/2, pi/2),random.uniform(-pi/2, pi/2))
+	goal.gripper_type = "sdhx"
 
 	#goal.grasp_database = "KIT"
 	goal.grasp_database = "OpenRAVE"
